requirements.txt
asgiref==3.8.1
boto3==1.36.19
botocore==1.36.26
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
jmespath==1.0.1
mysqlclient==2.2.7
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
pip-chill==1.0.3
psycopg2-binary==2.9.10
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
